refactor sections in settings
